# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lumina Auth API
  description: |
    Authentication service for the Lumina application.

    Provides user registration, JWT-based login, token refresh,
    and profile management. All authenticated endpoints expect a
    Bearer token in the `Authorization` header.

    ## Rate Limiting
    Rate-limited endpoints return standard `RateLimit-*` headers:
    - `RateLimit-Limit` — max requests allowed in the window
    - `RateLimit-Remaining` — remaining requests in the current window
    - `RateLimit-Reset` — seconds until the window resets

    | Scope | Window | Max Requests |
    |-------|--------|--------------|
    | General (all routes) | 15 min | 100 |
    | Auth (login / register) | 15 min | 5 |

    ## Client Code Generation
    This spec is designed for automatic typed-client generation.
    Recommended tools:
    - **Orval** — `npx orval --input ./docs/openapi.yaml`
    - **Kubb** — `npx kubb generate`
    - **Hey API** — `npx @hey-api/openapi-ts`
    - **openapi-typescript** — `npx openapi-typescript ./docs/openapi.yaml -o ./src/api.d.ts`
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh the access token
      description: |
        Exchanges a valid refresh token for a new access / refresh
        token pair. The previous refresh token should be discarded.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            example:
              refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '400':
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 'Refresh token is required'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 'Invalid or expired refresh token'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────────
  # Logout
  # ──────────────────────────────────────────────────
...
  /api/auth/profile:
    get:
      operationId: getProfile
      summary: Get the authenticated user's profile
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User no longer exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 'User not found'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────────
  # Me (alias)
  # ──────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token obtained from `/api/auth/login` or `/api/auth/register`.
        Pass it as `Authorization: Bearer <token>`.

  # ── Reusable Responses ───────────────────────────
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noToken:
              summary: No token provided
              value:
                success: false
                message: 'No authorization token provided'
            invalidToken:
              summary: Invalid or expired token
              value:
                success: false
                message: 'Invalid or expired token'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Internal server error'

  # ── Schemas ──────────────────────────────────────
  schemas:
...
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: The refresh token returned by login or register.

    # ── Domain Models ──
    UserProfile:
      type: object
      description: Public user representation (never contains the password hash).
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
          example: ['user']
      required: [id, email, username, createdAt, updatedAt, roles]

    # ── Responses ──
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
      required: [success, message]
...
    TokenRefreshResponse:
      type: object
      description: New token pair issued after a successful refresh.
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          type: object
          required: [accessToken, refreshToken]
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
      required: [success, message, data]
    ProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        message:
          type: string
        data:
          type: object
          required: [user]
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
      required: [success, message, data]